Move pyqir.generator to pyqir
