+++ pyqir/tests/test_external_functions.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-from pyqir.generator import (
+from pyqir import (
     BasicQisBuilder,
     SimpleModule,
     Type,
